Move messaging coordination into app and time modules
