+++ modules/CMakeLists.txt
@@ -15,6 +15,8 @@ message(STATUS "   TBX_VERB_LOGGING: ${TBX_VERB_LOGGING}")
 set(CMAKE_FOLDER "Modules")
 
 add_subdirectory(std)
+add_subdirectory(file_system)
+add_subdirectory(debugging)
 add_subdirectory(math)
 add_subdirectory(time)
 add_subdirectory(messaging)
@@ -24,6 +26,7 @@ add_subdirectory(app)
 set(CMAKE_FOLDER "Tests")
 
 add_subdirectory(std/tests)
+add_subdirectory(file_system/tests)
 add_subdirectory(time/tests)
 add_subdirectory(messaging/tests)
 add_subdirectory(plugin_api/tests)
