+++ modules/app/include/tbx/app/app_events.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "tbx/app/app_description.h"
-#include "tbx/messages/events/event.h"
+#include "tbx/messages/event.h"
 #include "tbx/time/delta_time.h"
 
 namespace tbx
