+++ modules/debugging/CMakeLists.txt
@@ -1,27 +1,23 @@
-tbx_add_module(PluginApi)
+tbx_add_module(Debugging)
 
 file(GLOB_RECURSE INCL CONFIGURE_DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/include/*.*")
 file(GLOB_RECURSE SRCS CONFIGURE_DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/src/*.*")
-target_sources(PluginApi PRIVATE ${SRCS} ${INCL})
+target_sources(Debugging PRIVATE ${SRCS} ${INCL})
 
-target_include_directories(PluginApi
+target_include_directories(Debugging
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
         $<INSTALL_INTERFACE:include>
     PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}
         ${CMAKE_CURRENT_SOURCE_DIR}/src
-        ${CMAKE_SOURCE_DIR}/modules/app/include
 )
 
-target_link_libraries(PluginApi
+target_link_libraries(Debugging
     PUBLIC
-        Tbx::Messaging
         Tbx::Std
+        Tbx::Messaging
         Tbx::FileSystem
-        Tbx::Debugging
-    PRIVATE
-        nlohmann_json
 )
 
-add_library(Tbx::PluginApi ALIAS PluginApi)
+add_library(Tbx::Debugging ALIAS Debugging)
