+++ modules/debugging/include/tbx/debugging/macros.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "tbx/debug/logging.h"
+#include "tbx/debugging/logging.h"
 
 #define TBX_TRACE_INFO(msg, ...)                                                                   \
     ::tbx::trace(::tbx::LogLevel::Info, __FILE__, __LINE__, msg __VA_OPT__(, ) __VA_ARGS__)
