+++ modules/std/CMakeLists.txt
@@ -2,6 +2,11 @@ tbx_add_module(Std)
 
 file(GLOB_RECURSE INCL CONFIGURE_DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/include/*.*")
 file(GLOB_RECURSE SRCS CONFIGURE_DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/src/*.*")
+
+list(FILTER INCL EXCLUDE REGEX "/file_system/")
+list(FILTER SRCS EXCLUDE REGEX "/file_system/")
+list(FILTER SRCS EXCLUDE REGEX "/string_path_operations\\.cpp$")
+
 target_sources(Std PRIVATE ${SRCS} ${INCL})
 
 target_include_directories(Std
