+++ modules/std/include/tbx/std/array.h
@@ -46,4 +46,4 @@ namespace tbx
     };
 }
 
-#include "../../../src/array.inl"
+#include "../../../src/std/array.inl"
