+++ modules/std/include/tbx/std/casting.h
@@ -59,4 +59,4 @@ namespace tbx
     bool try_as(Variant& value, TTo*& out);
 }
 
-#include "../../../src/casting.inl"
+#include "../../../src/std/casting.inl"
