+++ modules/std/include/tbx/std/list.h
@@ -89,4 +89,4 @@ namespace tbx
     };
 }
 
-#include "../../../src/list.inl"
+#include "../../../src/std/list.inl"
