+++ modules/std/include/tbx/std/smart_pointers.h
@@ -144,4 +144,4 @@ namespace tbx
     };
 }
 
-#include "../../../src/smart_pointers.inl"
+#include "../../../src/std/smart_pointers.inl"
