+++ modules/std/include/tbx/std/variant.h
@@ -47,4 +47,4 @@ namespace tbx
     };
 }
 
-#include "../../../src/variant.inl"
+#include "../../../src/std/variant.inl"
