+++ plugins/spdconsolelogger/src/spd_console_logger_plugin.cpp
@@ -1,6 +1,6 @@
 #include "spd_console_logger_plugin.h"
 #include "tbx/file_system/string_path_operations.h"
-#include "tbx/messages/commands/log_commands.h"
+#include "tbx/messages/log_commands.h"
 #include "tbx/std/casting.h"
 #include "tbx/std/smart_pointers.h"
 #include <spdlog/sinks/stdout_color_sinks.h>
